Separated spec parsing from generation
